LayoutTests/fast/dom/Document/document-open-removes-all-listeners-expected.txt

Event listeners are to be removed with immediate effect by document.open()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS once is true
PASS successfullyParsed is true

TEST COMPLETE
